requirements.txt
pywin32
# pywinauto
qiniu
Pillow
apscheduler
pyinstaller
PyYAML
wmi
pydes
# pycrypto
# pycryptodome
cryptography